refactor: rebrand from ai-test-lab to fm-app-toolkit
BREAKING CHANGE: Package renamed from ai_test_lab to fm_app_toolkit

- Update project name to FM App Toolkit (Foundation Model Application Toolkit)
- Comprehensive README rewrite emphasizing production patterns and business logic integration
- Package renamed from ai_test_lab to fm_app_toolkit across all files
- Updated mission: concrete implementation and testing patterns for production FM apps
- Position as complete toolkit, not just testing strategies
- Focus on bringing business logic closer to reality using LlamaIndex

Changes:
- Renamed main package directory from ai_test_lab/ to fm_app_toolkit/
- Updated all imports throughout the codebase
- New README with clear value proposition and production patterns
- Updated pyproject.toml with new name and description
- Updated Makefile references
- All 106 tests passing with new naming
