+++ fm_app_toolkit/agents/react_prompt_template.md
