+++ fm_app_toolkit/testing/mocks.py
@@ -4,9 +4,9 @@ This module re-exports all mock classes from their new locations
 to maintain backward compatibility with existing test imports.
 
 All new code should import directly from the specific modules:
-- ai_test_lab.testing.mock_chain for MockLLMWithChain
-- ai_test_lab.testing.mock_echo for MockLLMEchoStream  
-- ai_test_lab.testing.mock_rule_based for RuleBasedMockLLM
+- fm_app_toolkit.testing.mock_chain for MockLLMWithChain
+- fm_app_toolkit.testing.mock_echo for MockLLMEchoStream  
+- fm_app_toolkit.testing.mock_rule_based for RuleBasedMockLLM
 """
 
 from .mock_chain import MockLLMWithChain
